+++ src/controls/fields/fieldNameRenderer/index.ts
@@ -1,0 +1,2 @@
+export * from './FieldNameRenderer.module.scss';
+export * from './FieldNameRenderer';