+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Releases
 
+## 1.2.0
+
+**New controls**
+
+- Field controls are added to the project
+- `IFrameDialog` was added to the project
+
+**Fixes**
+
+- Fixed theming in the `WebPartTitle` control
+
 ## 1.1.3
 
 - `FileTypeIcon` icon fixed. This control should now also work in SPFx extensions.